HW-7: move homework files to dedicated folder
