+++ hw07-patterns/src/main/java/com/patterns/listener/Listener.java
@@ -6,8 +6,4 @@ import com.patterns.model.Message;
 public interface Listener {
 
     void onUpdated(Message msg);
-
-    // todo: 4. Сделать Listener для ведения истории (подумайте, как сделать, чтобы сообщения не портились)
-    // Уже есть заготовка - класс HistoryListener, надо сделать его реализацию
-    // Для него уже есть тест, убедитесь, что тест проходит
 }
